Fix some things in link fixer with complex flows. Add "Fix in place" in workflow load check.
