+++ web/html/links.html
@@ -119,7 +119,7 @@
     </div>
 
     <script type="module">
-      import {LinkPage} from './links.js';
+      import {LinkPage} from './link_page.js';
       new LinkPage();
     </script>
   </body>
